personalized-news-summarizer/requirements.txt
flask
markdown
newspaper3k
requests
bs4
Frozen-Flask
functions-framework==3.*
google-cloud-storage
pymongo
